refactor(testing): Move common testing utilities from tests
Moved common generator utility functions to testing module
so that it can be accessed anywhere to create fixtures.

Additionally, set up rectangularize_data tests.
